[CHORE] AwardService 추상화
